requirements.txt
libsass==0.13.7
pdfkit==0.6.1
pkg-resources==0.0.0
six==1.11.0
